src/legacy_to_port/sanali209/requirements.txt
flet~=0.21.2
loguru~=0.7.2
networkx~=3.3
neo4j~=5.18.0
tqdm~=4.65.2
annoy~=1.17.3
g4f~=0.3.1.9
langchain~=0.3.14
langchain-community~=0.3.14
kornia~=0.8.1
opencv-python~=4.9.0.80
numpy~=1.24.4
PyAutoGUI~=0.9.54
pynput~=1.7.7
torch~=2.4.1+cu121
clip~=1.0
pillow~=10.2.0
pygame~=2.5.2
mss~=10.0.0
future~=0.18.3
scipy~=1.13.0
keyboard~=0.13.5
easyocr~=1.7.1
pyglet~=2.0.16
timm~=0.9.16
torchvision~=0.19.1+cu121
pymongo~=4.6.3
scikit-learn~=1.7.1
gradio_client~=1.7.2
langchain-core~=0.3.29
langchain-google-genai~=2.0.8
gradio~=4.40.0
onnxruntime~=1.17.3
huggingface-hub~=0.23.3
requests~=2.28.2
transformers~=4.41.2
openai~=1.59.7
chromadb~=0.5.0
dask~=2024.9.1
diskcache~=5.6.3
ultralytics~=8.2.16
supervision~=0.25.1
autodistill~=0.1.29
sqlmodel~=0.0.18
dependency-injector~=4.45.0
psutil~=5.9.8
six~=1.16.0
bs4~=0.0.2
beautifulsoup4~=4.12.3
html2image~=2.0.4.3
ray~=2.45.0
pydantic~=2.10.5
IPTCInfo3~=2.1.4
Jinja2~=3.1.4
pyparsing~=3.2.1
protobuf~=4.25.5
Whoosh~=2.7.4
nltk~=3.6.7
sympy~=1.13.1
grpcio~=1.67.1
distributed~=2024.9.0
Send2Trash~=1.8.3
playwright~=1.51.0
python-slugify~=8.0.4
fs~=2.4.16
pyperclip~=1.8.2
vosk~=0.3.45
PyAudio~=0.2.14
pyarrow~=16.1.0
datasets~=2.12.0
duckduckgo_search~=8.0.1
NodeGraphQtPy~=0.6.38.6
QtPy~=2.4.3
SpeechRecognition~=3.14.2
Flask~=3.1.1
matplotlib~=3.9.0
deepface~=0.0.91
pandas~=1.5.3
python-dotenv~=1.0.1
uvicorn~=0.26.0
fastapi~=0.109.2
pywin32~=308
SQLAlchemy~=2.0.30
selenium~=4.9.1
notion~=0.0.28
PySide6~=6.9.0
gradio_client~=1.7.2
joblib~=1.4.2
pyexiv2~=2.15.4
trueskill~=0.4.5
duckduckgo_search~=8.0.1
gradio_client~=1.7.2
duckduckgo_search~=8.0.1
gradio_client~=1.7.2
umap-learn~=0.5.9.post2
vedo~=2025.5.4
duckduckgo_search~=8.0.1